requirements.txt
streamlit
nltk
spacy
pyresparser
pdfminer.six
docx2txt
pandas
numpy
